+++ src/Containers/Auth/AuthScreen.js
@@ -1,10 +1,15 @@
 import React, { Component } from 'react'
-import { Platform, StyleSheet, Text, View, TextInput, Image } from 'react-native';
-import { Icon } from 'react-native-elements';
-import styles from '../Styles/styles';
-import AuthForm from '../Components/AuthForm';
+import { Platform, StyleSheet, Text, View, TextInput, Image, Dimensions } from 'react-native';
+import { Icon } from 'react-native-elements/src/index';
+import styles from '../../Styles/styles';
+import AuthForm from '../../Components/AuthForm';
 
 class AuthScreen extends Component {
+
+  static navigationOptions = {
+    title: 'Home',
+    headerTitleStyle : {width : Dimensions.get('window').width}
+  };
 
   //Redirection to Registration screen
   _signUpRedirection = () => {
