+++ src/Containers/Content/ProfileScreen.js
@@ -1,13 +1,8 @@
-import React, { Component } from 'react'
-import { Text, View } from 'react-native'
-import { Icon } from 'react-native-elements';
+import React, {Component} from 'react'
+import {View} from 'react-native'
 
 
 export default class ProfileScreen extends Component {
-
-    static navigationOptions = {
-    title: 'Profile'
-  };
 
     render() {
         return (
