--- src/Containers/RegistrationScreen.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-import {Platform, StyleSheet, Text, View} from 'react-native';
-
-
-export default class RegistrationScreen extends Component {
-  render() {
-    return (
-      <View>
-        <Text>RegistrationScreen</Text>
-      </View>
-    )
-  }
-}
